2_file_handling/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydicom>=2.4.0
Pillow>=10.0.0 